+++ Laberinto/src/Sucesor.java
@@ -2,14 +2,14 @@
 public class Sucesor {
 	String accion;
 	int[] estado;
-	int costo_move;
+	float costo_move;
 	
 	
 	public Sucesor() {
 		
 	}
 	
-	public Sucesor(String accion, int[] estado, int costo_move) {
+	public Sucesor(String accion, int[] estado, float costo_move) {
 		this.accion = accion;
 		this.estado = estado;
 		this.costo_move = costo_move;
@@ -27,10 +27,10 @@ public class Sucesor {
 	public void setEstado(int[] estado) {
 		this.estado = estado;
 	}
-	public int getCosto_move() {
+	public float getCosto_move() {
 		return costo_move;
 	}
-	public void setCosto_move(int costo_move) {
+	public void setCosto_move(float costo_move) {
 		this.costo_move = costo_move;
 	}
